Applying coupons using React

+++ backend/controllers/payment.controller.js
@@ -140,7 +140,7 @@ async function createStripeCoupon(discountPercentage) {
 }
 
 async function createNewCoupon(userId) {
-	await Coupon.findOneAndDelete({ userId });
+	await Coupon.findOneAndDelete({userId: userId});
 
 	const newCoupon = new Coupon({
 		code: 'GIFT' + Math.random().toString(36).substring(2, 8).toUpperCase(),
